requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
huggingface_hub>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
ipykernel>=6.29.0
